Modified

+++ src/main/java/com/spring/springdata/employee/service/EmployeeService.java
@@ -2,7 +2,9 @@ package com.spring.springdata.employee.service;
 
 import com.spring.springdata.employee.model.Employee;
 
+import java.math.BigDecimal;
 import java.util.List;
+
 
 public abstract class EmployeeService {
 
@@ -12,7 +14,7 @@ public abstract class EmployeeService {
 
     public abstract List<Employee> getAllEmployee();
 
-    public abstract void updateEmployee(int id, float newSalary);
+    public abstract void updateEmployee(int id, BigDecimal newSalary);
 
     public abstract void deleteEmployee(int id);
